Create README.md
The dataset information on prescriptions that will be used for this task was provided by National Health Service (NHS) which was issued in England, focusing specifically on Bolton. We will be using SQL queries to analyze the data and provide insights into the types of medication being prescribed, the organizations doing the prescribing, and the quantities prescribed.